+++ R/sjPlotInteractions.R
@@ -1062,14 +1062,26 @@ sjp.eff.int <- function(fit,
       # combine lists
       if (is.null(int.term)) {
         # re-compute effects
-        eff.tmp <- effects::allEffects(fit, xlevels = xl,  KR = p.kr,
-                                       confidence.level = dot.args[["ci.lvl"]], ...)
+        eff.tmp <- effects::allEffects(
+          fit,
+          xlevels = xl,
+          KR = p.kr,
+          confidence.level = dot.args[["ci.lvl"]],
+          ...
+        )
         # reset data frame
         intdf <- data.frame(eff.tmp[[intpos[i]]])
       } else {
         # re-compute effects
-        eff.tmp <- effects::effect(int.term, fit, xlevels = xl, KR = p.kr,
-                                   confidence.level = dot.args[["ci.lvl"]], ...)
+        eff.tmp <-
+          effects::effect(
+            int.term,
+            fit,
+            xlevels = xl,
+            KR = p.kr,
+            confidence.level = dot.args[["ci.lvl"]],
+            ...
+          )
         # reset data frame
         intdf <- data.frame(eff.tmp)
       }
@@ -1106,8 +1118,8 @@ sjp.eff.int <- function(fit,
     # make sure x is numeric
     intdf$x <- sjmisc::to_value(intdf$x, keep.labels = F)
     # get name of response, for axis title
-    yaxisname <- sjmisc::get_label(stats::model.frame(fit)[[response.name]], 
-                                   def.value = response.name)
+    yaxisname <-
+      sjmisc::get_label(stats::model.frame(fit)[[response.name]], def.value = response.name)
     # -----------------------------------------------------------
     # check if we have linear regression
     # -----------------------------------------------------------
@@ -1286,7 +1298,7 @@ sjp.eff.int <- function(fit,
     # ------------------------------------------------------------
     # start plot
     # ------------------------------------------------------------
-    baseplot <- ggplot(intdf, aes_string(x = "x", y = "y", colour = "grp"))
+    baseplot <- ggplot(intdf, aes_string(x = "x", y = "y", colour = "grp", linetype = "grp"))
     # ------------------------------------------------------------
     # confidence interval?
     # ------------------------------------------------------------
@@ -1301,7 +1313,7 @@ sjp.eff.int <- function(fit,
         # -------------------------------------------------
         if (jitter.ci) {
           baseplot <- baseplot +
-            geom_errorbar(aes_string(ymin = "conf.low", ymax = "conf.high", colour = "grp"),
+            geom_errorbar(aes_string(ymin = "conf.low", ymax = "conf.high", colour = "grp", linetype = "grp"),
                           width = dot.args[["eb.width"]], show.legend = FALSE, 
                           position = position_dodge(.2)) +
             geom_point(position = position_dodge(.2), shape = 16) +
@@ -1311,7 +1323,7 @@ sjp.eff.int <- function(fit,
           upperLim.x <- upperLim.x + .2
         } else {
           baseplot <- baseplot +
-            geom_errorbar(aes_string(ymin = "conf.low", ymax = "conf.high", colour = "grp"),
+            geom_errorbar(aes_string(ymin = "conf.low", ymax = "conf.high", colour = "grp", linetype = "grp"),
                           width = dot.args[["eb.width"]], show.legend = FALSE) +
             geom_point(shape = 16) +
             geom_line(size = geom.size)